Create NetworkSniffer.py
This is a Python-based network packet sniffer built with Scapy. It captures and analyzes network traffic on a specified interface for example a Wi-Fi or a LAN, displaying source/destination IP addresses and TCP/UDP port details. it is ideal for network debugging and traffic analysis in the field of Cyber sceurity.